Removed selectorDocument and SelectorImage. SelectorFile replaced both.
