+++ src/_locales/ru/messages.json
@@ -70,9 +70,7 @@
 	"selector_edit_discardInitialElements": {
 		"message": "Отменить начальные элементы, которые доступны до нажатия"
 	},
-	"selector_edit_download_image": { "message": "Сохранить картинку" },
 	"selector_edit_download_file": { "message": "Сохранить файл" },
-	"selector_edit_download_document": { "message": "Сохранить документ" },
 	"selector_edit_attribute_name": { "message": "Имя атрибута" },
 	"selector_edit_value": { "message": "Значение" },
 	"selector_edit_style_name": { "message": "Имя стиля" },
@@ -131,8 +129,6 @@
 	"SelectorText": { "message": "Текст" },
 	"SelectorLink": { "message": "Ссылка" },
 	"SelectorPopupLink": { "message": "Всплывающая ссылка" },
-	"SelectorImage": { "message": "Изображение" },
-	"SelectorDocument": { "message": "Документ" },
 	"SelectorTable": { "message": "Таблица" },
 	"SelectorElementAttribute": { "message": "Атрибут" },
 	"SelectorElementStyle": { "message": "Стиль" },
