+++ src/devtools/views/SelectorEdit.html
@@ -312,40 +312,6 @@
 			</div>
 		</div>
 	</div>
-
-	<!-- downloadImage -->
-	<div class="form-group feature feature-downloadImage">
-		<label
-			for="downloadImage"
-			class="col-sm-2 control-label"
-			data-i18n="selector_edit_download_image"
-		></label>
-		<div class="col-sm-8">
-			<div class="checkbox">
-				<label>
-					<input type="checkbox" id="downloadImage" name="downloadImage"
-					{{#selector.downloadImage}} checked="checked" {{/selector.downloadImage}} />
-				</label>
-			</div>
-		</div>
-	</div>
-	<!-- downloadDocument-->
-	<div class="form-group feature feature-downloadDocument">
-		<label
-			for="downloadDocument"
-			class="col-sm-2 control-label"
-			data-i18n="selector_edit_download_document"
-		></label>
-		<div class="col-sm-8">
-			<div class="checkbox">
-				<label>
-					<input type="checkbox" id="downloadDocument" name="downloadDocument"
-					{{#selector.downloadDocument}} checked="checked" {{/selector.downloadDocument}}
-					/>
-				</label>
-			</div>
-		</div>
-	</div>
 	<!-- downloadFile-->
 	<div class="form-group feature feature-downloadFile">
 		<label
