+++ src/scripts/Controller.js
@@ -40,12 +40,6 @@ export default class SitemapController {
 				type: 'SelectorPopupLink',
 			},
 			{
-				type: 'SelectorImage',
-			},
-			{
-				type: 'SelectorDocument',
-			},
-			{
 				type: 'SelectorFile',
 			},
 			{
@@ -1110,8 +1104,6 @@ export default class SitemapController {
 			':checked'
 		);
 		const multiple = $('#edit-selector [name=multiple]').is(':checked');
-		const downloadImage = $('#edit-selector [name=downloadImage]').is(':checked');
-		const downloadDocument = $('#edit-selector [name=downloadDocument]').is(':checked');
 		const downloadFile = $('#edit-selector [name=downloadFile]').is(':checked');
 		const clickPopup = $('#edit-selector [name=clickPopup]').is(':checked');
 		const delay = $('#edit-selector [name=delay]').val();
@@ -1163,8 +1155,6 @@ export default class SitemapController {
 			discardInitialElements,
 			type,
 			multiple,
-			downloadImage,
-			downloadDocument,
 			downloadFile,
 			clickPopup,
 			extractAttribute,
